Server/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
eventlet==0.33.0
PyJWT==2.3.0
python-socketio==5.4.0
Flask-Migrate==3.1.0
alembic==1.7.1 